-perbaikan penyimpanan file ke folder public - cek ketersedian file jika akan dihapus

+++ app/Http/Controllers/Informasi/EventController.php
@@ -31,9 +31,10 @@
 
 namespace App\Http\Controllers\Informasi;
 
-use App\Http\Controllers\Controller;
-use App\Http\Requests\EventRequest;
 use App\Models\Event;
+use App\Http\Requests\EventRequest;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\File;
 
 class EventController extends Controller
 {
@@ -61,6 +62,7 @@ class EventController extends Controller
             $input['status'] = 'OPEN';
             Event::create($input);
         } catch (\Exception $e) {
+            dd($e);
             report($e);
             return back()->withInput()->with('error', 'Simpan Event gagal!');
         }
@@ -84,10 +86,8 @@ class EventController extends Controller
             if ($request->hasFile('attachment')) {
                 $lampiran = $request->file('attachment');
                 $fileName = $lampiran->getClientOriginalName();
-                $path     = "storage/event/" . $event->id . '/';
-                $lampiran->move($path, $fileName);
-                unlink(base_path('public/' . $event->file_dokumen));
-
+                $path     = "event/" . $event->id . '/';
+                $lampiran->move(base_path('public/'.$path), $fileName);
                 $input['attachment'] = $path . $fileName;
             }
             $event->update($input);
@@ -103,7 +103,9 @@ class EventController extends Controller
     {
         try {
             if ($event->delete()) {
-                unlink(base_path('public/' . $event->file_dokumen));
+                if ($event->file_dokumen != null &&  File::exists(base_path('public/' . $event->file_dokumen))) {
+                    unlink(base_path('public/' . $event->file_dokumen));
+                }
             }
         } catch (\Exception $e) {
             report($e);
